+++ Love4AnimalsAPI/Interfaces/ICampaignService.cs
@@ -1,13 +1,12 @@
-using System;
-
 using Love4AnimalsAPI.Models;
+
 namespace Love4AnimalsAPI.Interfaces;
 
 public interface ICampaignService
 {
     Task<List<Campaign>> GetAllAsync();
-    Task<Campaign?> GetByIdAsync(int id);
+    Task<Campaign?> GetByIdAsync(long id);
     Task<Campaign> CreateAsync(Campaign campaign);
-    Task<bool> UpdateAsync(int id, Campaign campaign);
-    Task<bool> DeleteAsync(int id);
+    Task<bool> UpdateAsync(long id, Campaign campaign);
+    Task<bool> DeleteAsync(long id);
 }
